+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to
 - ⬆️(back) upgrade lxml and pypdf
 - ✨(back) set allow_smart_web_search to False for all users
 - ✨(back) make allow_conversation_analytics user setting readonly in admin
+- 🚸(front) switch to new conversation when creating a project
 
 ### Fixed